github-pr-mcp/requirements.txt
mcp>=0.9.0
PyGithub>=2.1.1
python-dotenv>=1.0.0